requirements.txt
# Core dependencies for Signature Verification System
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
opencv-python==4.8.0.74
scikit-image==0.21.0
scikit-learn==1.3.0
Pillow==10.0.0
Flask==2.3.2
Flask-CORS==4.0.0

# Additional utilities
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
scipy==1.11.1

# Development tools
jupyter==1.0.0
pytest==7.4.0